backend/requirements.txt
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf_extra_fields==3.7.0
djoser==2.1.0
gunicorn==20.1.0
requests==2.26.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
isort==5.13.2
PyJWT==2.3.0
reportlab
Django==3.2.16
django-cors-headers==3.13.0
